refactor 24.10.07 view 분리
